// src/app/models/master-model/master-model.ts
export interface masterModal {
  id: any;
  type: string;
}
export interface SettingModel {
  id: any;
  orderNumber?: any;
  type: string;
  isActive: string;
}
export interface VoucherTypeModal {
  voucherTypeId: string;
  name: string;
  code: string;
  description: string;
  isActive: boolean;
}